chore(calls): remove dead ConversationRelay-era TTS/STT modules and rename relay-* files
Fast-follow cleanup after the media-stream v2 migration (#37047):
- delete voice-quality.ts, elevenlabs-client.ts, stt-hints.ts (test-only importers)
- remove the native-twilio voice-spec registry and collapse the TtsProviderDefinition union
- drop the dormant calls.voice.hints config key
- rename relay-setup-router/relay-verification/relay-access-wait to call-setup-router/call-verification/access-request-wait
