+++ assistant/src/__tests__/call-setup-flow-invite.test.ts
@@ -11,7 +11,7 @@ import type {
 import type {
   SetupOutcome,
   SetupResolved,
-} from "../calls/relay-setup-router.js";
+} from "../calls/call-setup-router.js";
 import type { TrustContext } from "../daemon/trust-context.js";
 
 // ---------------------------------------------------------------------------
