+++ assistant/src/__tests__/call-setup-flow.test.ts
@@ -14,7 +14,7 @@ import type { MediaStreamOutput } from "../calls/media-stream-output.js";
 import type {
   SetupOutcome,
   SetupResolved,
-} from "../calls/relay-setup-router.js";
+} from "../calls/call-setup-router.js";
 import type { TrustContext } from "../daemon/trust-context.js";
 import type { TrustClass } from "../runtime/trust-class.js";
