+++ assistant/src/calls/call-verification.ts
